Cover 'non secret' flow in logger
